# src/docs/swagger.yaml
definitions:
...
  dto.ReqCreateUpdateGoal:
    properties:
      description:
        type: string
      hours_count:
        type: number
      id:
        type: integer
      name:
        type: string
      project_id:
        type: integer
      time_end:
        type: string
      time_start:
        type: string
    required:
    - hours_count
    - name
    - project_id
    - time_end
    - time_start
    type: object
...
  dto.RespGoal:
    properties:
      description:
        type: string
      hours_count:
        type: number
      id:
        type: integer
      name:
        type: string
      project_id:
        type: integer
      time_end:
        type: string
      time_start:
        type: string
      user_id:
        type: integer
    type: object
...
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  pkg.Response:
    properties:
      body: {}
    type: object
info:
  contact: {}
paths:
...
  /goal/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a goal
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: can't find goal with such id
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Delete a goal
      tags:
      - goal
    get:
      consumes:
      - application/json
      description: Get goal by id
      parameters:
      - description: Goal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success get goal
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespGoal'
              type: object
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Show a post
      tags:
      - goal
  /goal/create:
    post:
      consumes:
      - application/json
      description: Create goal
      parameters:
      - description: goal info
        in: body
        name: goal
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateUpdateGoal'
      produces:
      - application/json
      responses:
        "200":
          description: success update goal
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespGoal'
              type: object
        "400":
          description: bad req
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf or permission denied
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "422":
          description: unprocessable entity
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Create goal
      tags:
      - goal
  /goal/edit:
    post:
      consumes:
      - application/json
      description: Update a goal
      parameters:
      - description: goal info
        in: body
        name: goal
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateUpdateGoal'
      produces:
      - application/json
      responses:
        "200":
          description: success update goal
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespGoal'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf or permission denied
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "422":
          description: unprocessable entity
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Update a goal
      tags:
      - goal
  /goal/usr:
    get:
      description: Get user goals or get user goals for a day
      parameters:
      - description: day for events
        in: query
        name: day
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success get goals
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/dto.RespGoal'
                  type: array
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get user goals
      tags:
      - goal
...
swagger: "2.0"